@@ -5,20 +5,22 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.7.0
+  sdk: '>=3.2.3 <4.0.0'
 
 dependencies:
   bloc: ^9.0.0
   cached_network_image: ^3.3.1
   carousel_slider: ^5.0.0
   cloud_firestore: ^5.6.5
-  cupertino_icons: ^1.0.8
-  curved_navigation_bar: ^1.0.6
+  cupertino_icons: ^1.0.2
+  curved_navigation_bar: ^1.0.3
   dio: ^5.3.3
-  easy_localization: ^3.0.7+1
+  easy_localization: ^3.0.3
+  file_picker: ^9.2.3
   firebase_auth: ^5.5.1
   firebase_core: ^3.12.1
   firebase_messaging: ^15.2.4
+  fl_chart: ^0.71.0
   flutter:
     sdk: flutter
   flutter_animate: ^4.5.2
@@ -27,22 +29,28 @@ dependencies:
   flutter_launcher_icons: ^0.14.3
   flutter_localizations:
     sdk: flutter
-  flutter_screenutil: ^5.9.3
+  flutter_screenutil: ^5.9.0
   flutter_svg: ^2.0.9
   gap: ^3.0.1
   get_it: ^8.0.3
   image_picker: ^1.1.2
   lottie: ^3.3.1
+  open_file: ^3.3.2
+  path_provider: ^2.1.2
+  pdf: ^3.10.7
   pinput: ^5.0.1
   pretty_dio_logger: ^1.3.1
+  provider: ^6.1.4
+  qr_flutter: ^4.1.0
   retrofit: ^4.0.3
   retrofit_generator: ^9.1.9
   shared_preferences: ^2.2.3
   shimmer: ^3.0.0
   smooth_page_indicator: ^1.2.1
+
 # build_runner: ^2.4.7
 dev_dependencies:
-  flutter_lints: ^5.0.0
+  flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
   intl_utils: ^2.8.5
