@@ -6,7 +6,7 @@ import 'package:healthy_sync/core/themes/app_color.dart';
 import 'package:healthy_sync/core/themes/styles.dart';
 
 class CustomTextField extends StatelessWidget {
-  final TextEditingController controller;
+  final TextEditingController? controller;
   final EdgeInsetsGeometry? contentPadding;
   final String hintText;
   final IconData? icon;
@@ -22,13 +22,18 @@ class CustomTextField extends StatelessWidget {
   final ValueChanged<String>? onFieldSubmitted;
   final FocusNode? focusNode;
   final OutlineInputBorder? border;
-
+  final Color? iconColor;
   final Color? borderColor;
   final Color? fillColor;
-
+  final TextStyle? hintTextStyle;
+  final bool isEnabled;
+  final double? borderRadius;
+  final IconData? suffixIcon;
   const CustomTextField({
     super.key,
-    required this.controller,
+    this.hintTextStyle,
+    this.borderRadius,
+    this.controller,
     required this.hintText,
     this.icon,
     this.keyboardType = TextInputType.text,
@@ -46,24 +51,29 @@ class CustomTextField extends StatelessWidget {
     this.borderColor,
     this.fillColor,
     this.contentPadding,
+    this.isEnabled = true,
+    this.iconColor,
+    this.suffixIcon,
   });
 
   @override
   Widget build(BuildContext context) {
     return Container(
       decoration: BoxDecoration(
-         color: Colors.white,
-        borderRadius: BorderRadius.circular(16.r),
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(borderRadius ?? 16.r),
         boxShadow: [
           BoxShadow(
             color: AppColor.grey.withValues(alpha: 0.3),
-            spreadRadius: 2,
-            blurRadius: 4,
-            offset: Offset(0, 0),
+            // شادو خفيف جدًا
+            spreadRadius: 0.4, // نشر بسيط
+            blurRadius: 4.4, // ضبابية خفيفة
+            offset: Offset(0, 0.4), // ارتفاع بسيط للشادو
           ),
         ],
       ),
       child: TextFormField(
+        enabled: isEnabled,
         controller: controller,
         keyboardType: keyboardType,
         obscureText: isPassword && !isPasswordVisible,
@@ -72,48 +82,53 @@ class CustomTextField extends StatelessWidget {
         focusNode: focusNode,
         decoration: InputDecoration(
           isDense: true,
-          contentPadding:
-              contentPadding ??
+          contentPadding: contentPadding ??
               EdgeInsets.symmetric(horizontal: 20.sp, vertical: 18.h),
           filled: true,
           fillColor: fillColor ?? AppColor.white,
-          hintStyle: TextStyles.font12GreyW400,
-          border: border ?? borderStyle(borderColor ?? AppColor.border),
-          focusedBorder: borderStyle(borderColor ?? AppColor.border),
-          enabledBorder: borderStyle(borderColor ?? AppColor.transparent),
-          errorBorder: borderStyle(AppColor.red),
-          focusedErrorBorder: borderStyle(AppColor.red),
+          hintStyle: hintTextStyle ?? TextStyles.font12GreyW400,
+          border: border ??
+              borderStyle(borderColor ?? AppColor.border, borderRadius ?? 16.r),
+          focusedBorder: borderStyle(
+            borderColor ?? AppColor.border,
+            borderRadius ?? 16.r,
+          ),
+          enabledBorder: borderStyle(
+            borderColor ?? AppColor.transparent,
+            borderRadius ?? 16.r,
+          ),
+          errorBorder: borderStyle(AppColor.red, borderRadius ?? 16.r),
+          focusedErrorBorder: borderStyle(AppColor.red, borderRadius ?? 16.r),
           floatingLabelBehavior: floatingLabelBehavior,
           labelText: labelText,
-          labelStyle:
-              labelStyle ??
+          labelStyle: labelStyle ??
               TextStyle(
                 color: AppColor.black,
                 fontWeight: FontWeight.bold,
                 fontSize: 16.sp,
               ),
           hintText: hintText,
-      
-          suffixIcon:
-              isPassword
-                  ? GestureDetector(
-                    onTap: togglePasswordVisibility,
-                    child: Icon(
-                      !isPasswordVisible
-                          ? Icons.visibility_off
-                          : Icons.visibility,
-                      size: 20.sp,
-                      color: AppColor.grey,
-                    ),
-                  )
-                  : null,
+          prefixIcon: icon != null
+              ? Icon(icon, size: 25.sp, color: iconColor ?? AppColor.black)
+              : null,
+          suffixIcon: isPassword
+              ? GestureDetector(
+                  onTap: togglePasswordVisibility,
+                  child: Icon(
+                    !isPasswordVisible
+                        ? Icons.visibility_off
+                        : Icons.visibility,
+                    size: 20.sp,
+                    color: AppColor.grey,
+                  ),
+                )
+              : null,
           errorStyle: TextStyles.font12BlueW400.copyWith(color: AppColor.red),
         ),
         style: TextStyles.font12BlueW400.copyWith(
           fontSize: ResponsiveHelper.isMobile(context) ? 16.sp : 24.sp,
           color: AppColor.black,
         ),
-      
         validator: validator,
       ),
     );
