@@ -5,14 +5,10 @@ sealed class AppAssets {
   static String homeIcon = "$iconPath/home.svg";
   static String remind = "$iconPath/remind.svg";
   static String forgotPassword = "$imagePath/forgotpassword.png";
-  static String image1 = "$imagePath/image1.svg";
-  static String image2 = "$imagePath/image2.svg";
-  static String image3 = "$imagePath/image3.svg";
-  static String image4 = "$imagePath/image4.svg";
+
   static String verification = "$imagePath/Verification.png";
   static String translationPath = "assets/translations";
   static String testIcon = "$iconPath/tistIcon.svg";
-  static String chatBot = "$iconPath/chatbot.svg";
   static String radiologyCenters = "$iconPath/RadiologyCenters.svg";
   static String medicationTime = "$imagePath/pills-pill-svgrepo-com.png";
   static String loading = "$imagePath/loading.json";
@@ -25,4 +21,20 @@ sealed class AppAssets {
   static String loading3 = "$imagePath/loading3.json";
   static String loading4 = "$imagePath/loading4.json";
   static String welcome = "$imagePath/welcome.png";
+  static String chatSvg =
+      "$iconPath/chat_talk_conversation_sms_bubble_icon-icons.com_59983.svg";
+  static String injectionSvg = "$iconPath/injection.svg";
+  static String pillsSvg = "$iconPath/pills.svg";
+  static String ellipse = "$imagePath/Ellipse.svg";
+  static String doctor = "$imagePath/doctor.png";
+  static String patient = "$imagePath/patient.png";
+  static String rivet = "$imagePath/rivet.png";
+  static String syringe = "$imagePath/syringe.png";
+  static String syringe1 = "$imagePath/syringe1.png";
+  static String syringe2 = "$imagePath/syringe2.png";
+  static String tist = "$imagePath/tist.png";
+  static String carry = "$imagePath/carry.png";
+  static String button = "$imagePath/button.png";
+  static String male = "$iconPath/Male.svg";
+  static String female = "$iconPath/Female.svg";
 }
