@@ -1,22 +1,12 @@
 sealed class EndPoints {
-  static String baseUrl = "https://emr-system-production.up.railway.app/api/";
-  //static String baseUrl = "https://codingarabic.online/api";
+  static String baseUrl = "https://emr-system-henna.vercel.app/";
   static String auth = "auth";
   static String login = "$auth/login";
-  static String register = "$auth/register";
-  static String productsBestseller = "/products-bestseller";
-  static String sliders = "/sliders";
-  static String addToWishlist = "/add-to-wishlist";
-  static String wishlist = "/wishlist";
-  static String removeFromWishlist = "/remove-from-wishlist";
-  static String products = "/products";
-  static String cart = "/cart";
-  static String addToCart = "/add-to-cart";
-  static String updateCart = "/update-cart";
-  static String removeFromCart = "/remove-from-cart";
-  static String profile = "/profile";
-  static String updateProfile = "/update-profile";
-  static String updatePassword = "/update-password";
+  static String register = "$auth/signup";
 
-  static String updateUser = "/users";
+  static String profile = "profile";
+  static String updateProfile = "update-profile";
+  static String updatePassword = "update-password";
+
+  static String updateUser = "users";
 }
