@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:healthy_sync/core/constants/app_assets.dart';
 import 'package:healthy_sync/core/constants/enum.dart';
+import 'package:healthy_sync/core/themes/app_color.dart';
 import 'package:healthy_sync/core/themes/styles.dart';
 import 'package:healthy_sync/core/translations/locale_keys.g.dart';
 import 'package:healthy_sync/core/helpers/extensions.dart';
@@ -12,12 +13,21 @@ class SignUpScreen extends StatelessWidget {
   final UserType userType;
   const SignUpScreen({super.key, required this.userType});
   String handleUserType() {
-    return userType == UserType.doctor ? 'دكتور' : 'مريض';
+    return userType == UserType.doctor
+        ? LocaleKeys.doctor.tr()
+        : LocaleKeys.patient.tr();
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: AppBar(
+        backgroundColor: AppColor.white,
+        leading: IconButton(
+          onPressed: () => context.pop(),
+          icon: const Icon(Icons.arrow_back),
+        ),
+      ),
       body: SafeArea(
         child: SingleChildScrollView(
           child: Column(
