@@ -30,115 +30,118 @@ class _VerificationScreenState extends State<VerificationScreen> {
     var cubitOTB = context.read<VerificationCubit>();
     return Scaffold(
       appBar: AppBar(
+        toolbarHeight: 48.sp,
         backgroundColor: AppColor.white,
         iconTheme: IconThemeData(color: AppColor.black, size: 24.sp),
         centerTitle: true,
         title: Text(
           LocaleKeys.verification.tr(),
-          style: TextStyles.font12DarkBlueW400
+          style: TextStyles.font16DarkBlueW500.copyWith(
+            fontWeight: FontWeight.bold,
+          ),
         ),
       ),
       body: Padding(
         padding: EdgeInsets.all(20.sp),
-        child:
+        child: Column(
+          children: [
+            Image.asset(
+              AppAssets.verification,
+              height: 70.sp,
+              width: 70.w,
+              fit: BoxFit.cover,
+            ),
+            30.H,
+            Text(
+              LocaleKeys.enterTheCode.tr(),
+              style: TextStyles.font12DarkBlueW400,
+            ),
+            16.H,
+            Text.rich(
+              TextSpan(
+                text: LocaleKeys.CodeSentTo.tr(),
+                style: TextStyles.font12DarkBlueW400,
+                children: [
+                  TextSpan(
+                    text: widget.phone,
+                    style: TextStyles.font12DarkBlueW400.copyWith(
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            16.H,
             Column(
               children: [
-                Image.asset(
-                  AppAssets.verification,
-                  height: 70.sp,
-                  width: 70.w,
-                  fit: BoxFit.cover,
-                ),
-
-                30.H,
-                Text(
-                  LocaleKeys.enterTheCode.tr(),
-                  style: TextStyles.font12DarkBlueW400
-                ),
-                16.H,
-                Text.rich(
-                  TextSpan(
-                    text: LocaleKeys.CodeSentTo.tr(),
-                    style: TextStyles.font12DarkBlueW400,
-                    children: [
-                      TextSpan(
-                        text: widget.phone,
-                        style: TextStyles.font12DarkBlueW400
-                      ),
-                    ],
-                  ),
-                ),
-                16.H,
-                Column(
+                Row(
                   children: [
-                    Row(
-                      children: [
-                        Expanded(
-                          child: Text(
-                            LocaleKeys.Code.tr(),
-                            style:TextStyles.font12DarkBlueW400
-                          ),
-                        ),
-                        8.W,
-                        TimerWidget(),
-                      ],
-                    ),
-                    16.H,
-                    Pinput(
-                      showCursor: false,
-                      controller: cubitOTB.pinController,
-                      length: 4,
-                      defaultPinTheme: defaultPinTheme,
-                      focusedPinTheme: defaultPinTheme,
-                      submittedPinTheme: defaultPinTheme.copyWith(
-                        decoration: BoxDecoration(
-                          color: AppColor.secondary,
-                          border: Border.all(
-                            color: AppColor.secondary,
-                            width: 2,
-                          ),
-                          borderRadius: BorderRadius.circular(10),
-                        ),
+                    Expanded(
+                      child: Text(
+                        LocaleKeys.Code.tr(),
+                        style: TextStyles.font12DarkBlueW400,
                       ),
-                      onCompleted: (pin) {
-                        ScaffoldMessenger.of(context).showSnackBar(
-                          SnackBar(content: Text("Entered PIN: $pin")),
-                        );
-                        cubitOTB.isCodeEmptyShown = false;
-                      },
                     ),
-                    if (cubitOTB.isCodeEmptyShown)
-                      Column(
-                        children: [
-                          30.H,
-                          Text(
-                            cubitOTB.text ?? LocaleKeys.codeIsRequired.tr(),
-                            style: TextStyles.font12DarkBlueW400
-                          ),
-                        ],
-                      )
-                    else
-                      30.H,
-                    30.H,
-                    CustomButton(
-                      name:LocaleKeys.verify.tr(),
-                      onTap: () {
-                        setState(() {
-                          if (cubitOTB.pinController.text.length < 4) {
-                            cubitOTB.text = "❌ لازم تدخل 4 أرقام!";
-                            log("❌ لازم تدخل 4 أرقام!");
-                            cubitOTB.isCodeEmptyShown = true;
-                          } else {
-                            log("✅ الكود صحيح: ${cubitOTB.pinController.text}");
-                            cubitOTB.isCodeEmptyShown = false;
-                          }
-                        });
-                      },
-                    ),
+                    8.W,
+                    const TimerWidget(),
                   ],
                 ),
+                16.H,
+                Pinput(
+                  showCursor: false,
+                  controller: cubitOTB.pinController,
+                  length: 4,
+                  defaultPinTheme: defaultPinTheme,
+                  focusedPinTheme: defaultPinTheme,
+                  submittedPinTheme: defaultPinTheme.copyWith(
+                    decoration: BoxDecoration(
+                      color: AppColor.mainBlue,
+                      border: Border.all(
+                        color: AppColor.mainBlue,
+                        width: 2,
+                      ),
+                      borderRadius: BorderRadius.circular(10),
+                    ),
+                  ),
+                  onCompleted: (pin) {
+                    ScaffoldMessenger.of(context).showSnackBar(
+                      SnackBar(content: Text("Entered PIN: $pin")),
+                    );
+                    cubitOTB.isCodeEmptyShown = false;
+                  },
+                ),
+                if (cubitOTB.isCodeEmptyShown)
+                  Column(
+                    children: [
+                      30.H,
+                      Text(
+                        cubitOTB.text ?? LocaleKeys.codeIsRequired.tr(),
+                        style: TextStyles.font12DarkBlueW400,
+                      ),
+                    ],
+                  )
+                else
+                  30.H,
+                30.H,
+                CustomButton(
+                  name: LocaleKeys.verify.tr(),
+                  onTap: () {
+                    setState(() {
+                      if (cubitOTB.pinController.text.length < 4) {
+                        cubitOTB.text = "❌ لازم تدخل 4 أرقام!";
+                        log("❌ لازم تدخل 4 أرقام!");
+                        cubitOTB.isCodeEmptyShown = true;
+                      } else {
+                        log("✅ الكود صحيح: ${cubitOTB.pinController.text}");
+                        cubitOTB.isCodeEmptyShown = false;
+                      }
+                    });
+                  },
+                ),
               ],
-            ).scrollable(),
+            ),
+          ],
+        ).scrollable(),
       ),
     );
   }
