@@ -3,12 +3,14 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:healthy_sync/core/service/local/shared_helper.dart';
+import 'package:flutter/services.dart';
 
 import 'package:healthy_sync/healthy_sync.dart';
 import 'package:healthy_sync/core/constants/app_assets.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
+  await SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersiveSticky);
   // await Firebase.initializeApp();
   await SharedHelper.init();
   // SharedHelper.clear();
@@ -17,9 +19,8 @@ void main() async {
   runApp(
     EasyLocalization(
       supportedLocales: [Locale('en'), Locale('ar')],
-      path:
-          AppAssets
-              .translationPath, // <-- change the path of the translation files
+      path: AppAssets
+          .translationPath, // <-- change the path of the translation files
       fallbackLocale: Locale('en'),
       startLocale: Locale("ar"),
 
